added timer rendering, cpu usage jumps up to 0.7%, memory 1%, maybe use itoa instead of sprintf?

+++ timer.h
@@ -5,7 +5,7 @@
 
 typedef struct {
 	char *activityName;
-	double totalTime;
+	int totalTime;
 	time_t startStamp;
 } Activity;
